backend/requirements2.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
Pillow
numpy
pydantic
python-dotenv
httpx